+++ src/components/molecules/popover/index.ts
@@ -1,10 +1,5 @@
 export { PopoverWrapper } from "@components/molecules/popover/popover";
-export { PopoverCloseButton } from "@components/molecules/popover/popoverCloseButton";
-export { PopoverContent, PopoverClose, PopoverTrigger } from "@components/molecules/popover/popoverContent";
+export { PopoverContent, PopoverTrigger } from "@components/molecules/popover/popoverContent";
 export { PopoverContentBase } from "@components/molecules/popover/popoverContentBase";
-export { PopoverList } from "@components/molecules/popover/popoverList";
-export {
-	PopoverListContent,
-	PopoverListClose,
-	PopoverListTrigger,
-} from "@components/molecules/popover/popoverListContent";
+export { PopoverListWrapper } from "@components/molecules/popover/popoverList";
+export { PopoverListContent, PopoverListTrigger } from "@components/molecules/popover/popoverListContent";
