+++ src/components/molecules/popover/popoverContent.tsx
@@ -1,6 +1,5 @@
 import * as React from "react";
 
-import { PopoverCloseButton } from "./popoverCloseButton";
 import { PopoverContentBase } from "./popoverContentBase";
 import { usePopoverContext } from "@contexts/usePopover";
 import { PopoverTriggerProps } from "@src/interfaces/components";
@@ -49,5 +48,3 @@ export const PopoverContent = React.forwardRef<HTMLDivElement, React.HTMLProps<H
 		/>
 	);
 });
-
-export const PopoverClose = PopoverCloseButton;
