+++ src/components/molecules/popover/popoverTrigger.tsx
@@ -1,0 +1,28 @@
+import React, { forwardRef, isValidElement } from "react";
+
+import { useMergeRefs } from "@floating-ui/react";
+
+import { usePopoverContext } from "@contexts";
+import { PopoverTriggerProps } from "@src/interfaces/components";
+
+export const PopoverTrigger = forwardRef<HTMLElement, React.HTMLProps<HTMLElement> & PopoverTriggerProps>(
+	function PopoverTrigger({ children, ...props }, propRef) {
+		const context = usePopoverContext();
+		const childrenRef = isValidElement(children) ? (children as any).ref : null;
+		const ref = useMergeRefs([context.refs.setReference, propRef, childrenRef].filter(Boolean));
+
+		return (
+			<button
+				data-state={context.open ? "open" : "closed"}
+				onClick={() => {
+					context.setOpen(!context.open);
+				}}
+				ref={ref}
+				type="button"
+				{...context.getReferenceProps(props)}
+			>
+				{children}
+			</button>
+		);
+	}
+);
